requirements_brief_django2.0.10 - 1.txt
future==0.17.1
httplib2==0.12.0
MarkupSafe==1.1.0
Pygments==2.3.1
pyparsing==2.3.1
pyzmq==17.1.2
# mysqlclient==1.4.1
requests==2.21.0
#Django==2.0.5
django-reversion==3.0.3
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-formtools==2.1
XlsxWriter==1.1.2
django-import-export==1.0.0
Markdown==3.0.1
Pillow==5.4.1
django-markdownx==2.0.28
django-pure-pagination==0.3.0
django-simple-captcha==0.5.6
# restful framework
coreapi==2.3.3
markdown==3.0.1
djangorestframework==3.9.1
django-filter==2.1.0
django-guardian==1.5.0
django-cors-headers==2.4.0
djangorestframework-jwt==1.11.0
